refactor(Textfield): delete interface.tsx and define component in index.tsx
