+++ src/components/inputs/Textfield/index.tsx
@@ -1,7 +1,27 @@
 import { useState } from "react";
+import { MdOutlineError, MdCheckCircle } from "react-icons/md";
 
-import { TextfieldUI } from "./interface";
+import { Label } from "@inputs/Label";
+import { Text } from "@data/Text";
+
 import { InputType, Size, State } from "./props";
+
+import {
+  StyledContainer,
+  StyledContainerLabel,
+  StyledInputContainer,
+  StyledInput,
+  StyledIcon,
+  StyledErrorMessageContainer,
+  StyledValidMessageContainer,
+} from "./styles";
+
+export interface IMessageProps {
+  state?: State;
+  disabled?: boolean;
+  errorMessage?: string;
+  validMessage?: string;
+}
 
 export interface ITextfieldProps {
   label?: string;
@@ -25,6 +45,51 @@ export interface ITextfieldProps {
   readOnly?: boolean;
   focused?: boolean;
 }
+
+const getTypo = (size: Size) => {
+  if (size === "compact") {
+    return "medium";
+  }
+  return "large";
+};
+
+const Invalid = (props: IMessageProps) => {
+  const { disabled, state, errorMessage } = props;
+
+  return (
+    <StyledErrorMessageContainer disabled={disabled} state={state}>
+      <MdOutlineError />
+      <Text
+        type="body"
+        size="small"
+        margin="8px 0px 0px 4px"
+        appearance="error"
+        disabled={disabled}
+      >
+        {errorMessage && `(${errorMessage})`}
+      </Text>
+    </StyledErrorMessageContainer>
+  );
+};
+
+const Success = (props: IMessageProps) => {
+  const { disabled, state, validMessage } = props;
+
+  return (
+    <StyledValidMessageContainer disabled={disabled} state={state}>
+      <MdCheckCircle />
+      <Text
+        type="body"
+        size="small"
+        margin="8px 0px 0px 4px"
+        appearance="success"
+        disabled={disabled}
+      >
+        {validMessage}
+      </Text>
+    </StyledValidMessageContainer>
+  );
+};
 
 const Textfield = (props: ITextfieldProps) => {
   const {
@@ -68,28 +133,88 @@ const Textfield = (props: ITextfieldProps) => {
   };
 
   return (
-    <TextfieldUI
-      label={label}
-      name={name}
-      id={id}
-      placeholder={placeholder}
-      disabled={disabled}
-      type={type}
-      value={value}
-      onChange={onChange}
-      iconBefore={iconBefore}
-      iconAfter={iconAfter}
-      required={required}
-      size={size}
-      state={state}
-      errorMessage={errorMessage}
-      validMessage={validMessage}
-      fullwidth={fullwidth}
-      focused={focused}
-      onFocus={interceptFocus}
-      onBlur={interceptBlur}
-      readOnly={readOnly}
-    />
+    <StyledContainer fullwidth={fullwidth} disabled={disabled}>
+      <StyledContainerLabel
+        alignItems="center"
+        wrap="wrap"
+        size={size}
+        disabled={disabled}
+      >
+        {label && (
+          <Label
+            htmlFor={id}
+            disabled={disabled}
+            focused={focused}
+            invalid={state === "invalid" ? true : false}
+            size={getTypo(size!)}
+          >
+            {label}
+          </Label>
+        )}
+
+        {required && !disabled && (
+          <Text type="body" size="small" appearance="dark">
+            (Required)
+          </Text>
+        )}
+      </StyledContainerLabel>
+
+      <StyledInputContainer
+        disabled={disabled}
+        focused={focused}
+        state={state}
+        iconBefore={iconBefore}
+        iconAfter={iconAfter}
+      >
+        {iconBefore && (
+          <StyledIcon disabled={disabled} iconBefore={iconBefore}>
+            {iconBefore}
+          </StyledIcon>
+        )}
+
+        <StyledInput
+          label={label}
+          name={name}
+          id={id}
+          placeholder={placeholder}
+          disabled={disabled}
+          type={type}
+          value={value}
+          iconBefore={iconBefore}
+          iconAfter={iconAfter}
+          required={required}
+          size={size}
+          state={state}
+          fullwidth={fullwidth}
+          focused={focused}
+          onChange={onChange}
+          onFocus={interceptFocus}
+          onBlur={interceptBlur}
+          readOnly={readOnly}
+        />
+
+        {iconAfter && (
+          <StyledIcon iconAfter={iconAfter} disabled={disabled}>
+            {iconAfter}
+          </StyledIcon>
+        )}
+      </StyledInputContainer>
+
+      {state === "invalid" && (
+        <Invalid
+          disabled={disabled}
+          state={state}
+          errorMessage={errorMessage}
+        />
+      )}
+      {state === "valid" && (
+        <Success
+          disabled={disabled}
+          state={state}
+          validMessage={validMessage}
+        />
+      )}
+    </StyledContainer>
   );
 };
